EfficientNet_Lite2/requirements.txt
numpy
onnx
onnxruntime
Pillow
tqdm
